read data from mintManagerFunctions
