+++ frontend/src/libs/mintManagerFunctions.ts
@@ -1,13 +1,12 @@
 import contract from "../contracts/MintNFT.json";
-import { ethers } from "ethers";
+import { BigNumber, ethers } from "ethers";
 import { ipfs2http } from "../../utils/ipfs2http";
 import axios from "axios";
-import { IOwnedNFT } from "../hooks/useMintNFTManager";
+import { INFTMetaData, IOwnedNFT } from "../hooks/useMintNFTManager";
 const provierRpc = process.env.NEXT_PUBLIC_PROVIDER_RPC!;
 const contractAddress = process.env.NEXT_PUBLIC_CONTRACT_MINT_NFT_MANAGER!;
 
 const getMintNFTManagerContract = () => {
-  console.log('*****getRPC:', provierRpc)
   const provider = new ethers.providers.JsonRpcProvider(provierRpc);
   const _contract = new ethers.Contract(
     contractAddress,
@@ -30,4 +29,13 @@ export const getOwnedNFTsFromAddress = async (address: string) => {
     metadata.push({ tokenId: tokenId, metaData: data });
   }
   return metadata;
+}
+export const getNFTDataFromAddress = async (tokenId: BigNumber) => {
+  const mintNFTManager = getMintNFTManagerContract();
+  if (!mintNFTManager)
+    throw new Error("Cannot find mintNFTManager contract");
+  const tokenURI = await mintNFTManager.tokenURI(tokenId);
+  const path = ipfs2http(tokenURI);
+  const { data } = await axios.get(path);
+  return data as INFTMetaData;
 }
